requirements.txt
pyyaml
cx-Freeze
paramiko
pycrypto
sarge
